disconnect session selector from game
