+++ src/features/selector/Selector.tsx
@@ -1,10 +1,11 @@
 import * as React from "react";
-import { ChangeEvent, useEffect, useState } from "react";
+import { useState } from "react";
 import Session from "../session/Session";
-import { Box, Button, Grid, Stack, TextField, Typography } from "@mui/material";
+import { Grid, Typography } from "@mui/material";
 import { setSessionId } from "./selectorSlice";
 import { useAppDispatch, useAppSelector } from "../../hooks";
 import { useNewSessionQuery } from "../../services/session";
+import SessionConnector from "../../components/session_connector/SessionConnector";
 
 export default function Selector() {
   // generate indicates if a new session ID should be generated
@@ -12,8 +13,6 @@ export default function Selector() {
 
   // ready indicates if the game is ready to be rendered (e.g. any session ID has been selected)
   const [ready, setReady] = useState<boolean>(false);
-
-  let sessionIdInputRef;
 
   // prefill value if set from generation (but component should not be controlled)
   const selectorState = useAppSelector((state) => state.selector);
@@ -31,17 +30,10 @@ export default function Selector() {
     setGenerate(false);
   }
 
-  useEffect(() => {
-    if (sessionId && sessionIdInputRef) {
-      sessionIdInputRef.value = sessionId;
-    }
-  });
-
-  function connectSession(e: ChangeEvent<HTMLFormElement>) {
-    e.preventDefault();
+  function connectSession(selectedSessionId: string) {
     setReady(true);
     setGenerate(false);
-    dispatch(setSessionId(e.target["session-id"].value));
+    dispatch(setSessionId(selectedSessionId));
   }
 
   function createSession() {
@@ -52,50 +44,11 @@ export default function Selector() {
   function content() {
     if (isLoading || !ready) {
       return (
-        <Grid container justifyContent="center" alignItems="center">
-          <Grid item xs={3}>
-            <Box component="form" onSubmit={connectSession} autoComplete="off">
-              <Stack spacing={2}>
-                <TextField
-                  label="Username"
-                  id="username"
-                  sx={{
-                    width: "100%",
-                  }}
-                />
-                <TextField
-                  label="Session ID"
-                  id="session-id"
-                  inputRef={(e) => (sessionIdInputRef = e)}
-                  sx={{
-                    width: "100%",
-                  }}
-                />
-                <Grid container>
-                  <Grid item xs={6} sx={{ pr: 1 }}>
-                    <Button
-                      variant="contained"
-                      type="submit"
-                      sx={{ width: "100%" }}
-                    >
-                      Connect
-                    </Button>
-                  </Grid>
-                  <Grid item xs={6} sx={{ pl: 1 }}>
-                    <Button
-                      variant="contained"
-                      color="secondary"
-                      onClick={createSession}
-                      sx={{ width: "100%" }}
-                    >
-                      Create New
-                    </Button>
-                  </Grid>
-                </Grid>
-              </Stack>
-            </Box>
-          </Grid>
-        </Grid>
+        <SessionConnector
+          connect={connectSession}
+          createSession={createSession}
+          sessionId={sessionId}
+        />
       );
     } else {
       return <Session loading={false} sessionId={sessionId} />;
